Added reconnect logic.
By default it's disabled and for each sequence attempt the normal NetStatus even will be fired. The setting to adjust the reconnect attempts is called "reconnect_count_limit". The reconnect attempt uses the same connection string and parameters initially passed into the class.

I also standardized on the variable names so private begin with an underscore.

As always if you find any bugs in this implementation please let me know.