Fixed a race condition with reject messages coming from the server.
With the advent of the GUID being passed we now have a potential race condition serverside where the return value of a reject connection will happen in seemingly random orders. This poses an issue when you need to catch a custom param in the reject connection especially if each connection is not treated equally. So we now catch the first reject connection value and return that provided all the other connections fail. 

We've also moved away from a true shotgun approach seeing as a sequential connect at a low rate works equally well and consolidates the code down. If shotgun connect is set to true we simply lower the connection rate to 100ms which gives us a buffer of 300ms prior to the first rtmpt connection attempt after the first rtmp connect attempt. If more time is required simply disable the shotgun connect and increase the connection rate.