Stability improvement for mobile and resource improvement
When using this with mobile IOS Air apps if you connect in bulk the
application stability will suffer. I think this is due to the GC
implementation. It can cause the app to crash. To fix this enable
sequential_connect so that it will try a connection one at a time and if
a failure occurs in the form of a Reject or Fail then it will try the
next port in the sequence.

The class also no longer front loads all the PortConnections but rather
creates them on-demand so that only the needed PortConnections exist or
are created.