Reconnect fixes for sequential_connect and other goodies.
When sequential connecting and trying to utilize the reconnect logic the
sequence would fire once.. and then bail. We now clear out the ncTypes
array with each "connect" call into the class.

Also put in a timer to limit the wait time of a port connection to 30
seconds. This fixed a problem with an AIR IOS app trying to connect to a
restarting wowza box and it would hang on port 80 indefinitely.