Added a port test and added rtmfp to connection pool.
Originally if rtmfp was specified it would ignore rtmp and rtmpt. This
has been fixed such that it prioritizes rtmfp to be tried first.. if
that fails then rtmp and rtmpt are tried. This ensures that the best
protocol is attempted first.

Added the ability to put the class into port test mode to scan which
port and protocol to use when connecting.