Update to support smarter proxy resolution.
If proxy is set to none then we will assume that if a normal rtmp connection fails we want to try CONNECT through a proxy for performance reasons then ultimately fail back to rtmpt. If a proxyType is set then this will be skipped and then the user provide proxyType will be used.

Added a few other tests as well.