Fixed up a crash that was occurring when used behind a proxy that wouldn't let go of a pending connection.
We have our own internal timeout but when trying to close the pending connection it would put the player into a state of hungness.

Also fixed up the port test update to not fire on every port iteration but once after they've all been updated... no need to be so chatty.

Added a test to ensure we don't expose a crash/error again if close is called on an inactive netconnection.