Slight fix with response time reporting.
Prior to this fix we were reporting on the internal connection time of the PortConnection and not the overall time it took for NCS to establish a connection with port failbacks. This fixes that so that the time reported by NCS is the aggregate time it took to establish a connection. To access the time the internal PortConnection took to establish a connection for a specified port/protocol it would just be a matter of querying the library for the current netconnection and asking for the response time for it... like so:

var internal_connection_time:Number = ncs.connection.response_time;

VS getting the overall time to connect

var connection_time:Number = ncs.response_time;